added: option to add menu items for url-based quicklinks (not entities) chore: plugin code cleanup

+++ actions/reorder.php
@@ -1,6 +1,6 @@
 <?php
 
-$container_guid = get_input("container_guid", elgg_get_logged_in_user_guid());
+$container_guid = (int) get_input("container_guid", elgg_get_logged_in_user_guid());
 $ordered_guids = get_input("guids", "");
 
 if (get_entity($container_guid)) {
